==> 01-import.Rmd <==
---
title: "Import food desert info"
output:
  html_document:
    df_print: paged
knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, encoding = encoding, output_dir = "docs") })
---

```{r setup, echo=T, results='hide', message=F, warning=F}
library(tidyverse)
library(janitor)
library(readxl)
```

## Import the data

I'm starting from the original downloaded data in `data-raw`.

```{r}
fa_raw <- read_excel("data-raw/DataDownload2015.xlsx", sheet = "Food Access Research Atlas")
```

Look at names so we can select them:

```{r}
fa_raw %>% names()
```

### Select and filter data

- This list based on earlier csvkit work. This could change.
- Filtering to Travis County, Texas

```{r}
fa_travis <- fa_raw %>%
  select(c(1:6,10:13,15:24)) %>% # selects columns by position
  filter(State == "Texas", County == "Travis")
```

```{r}
fa_travis %>% write_rds("data-processed/fa_travis.rds")
```
