==> 02-data-wrangle.Rmd <==
---
title: "Data wrangling"
output:
  html_document:
    df_print: paged
knit: (function(inputFile, encoding) { rmarkdown::render(inputFile, encoding = encoding, output_dir = "docs") })
---

## The plan

- We are starting with Travis County data, which was created in [01-import.rmd](01-import.Rmd).
- We work on the LILA (Low Income, Low Access) columns first, so we drop the LA columns.
- We create a LA dataset to work on later.
- Reorder the columns to put LILA at the end, in order. This is just for clarity.
- Create a new `Limited_Income_Limited_Access` column based on our logic noted below.
- We do a similar function for the LA dataset to get `Limited_Access`.
- We create our final dataset by joining the new `Limited_Access` column to the LILA dataset.
- Once this is all done, we could export to join with the shapefile in QGIS.
- I also mapped the results in R in the next workbook.

## Libraries

```{r setup, echo=T, results='hide', message=F, warning=F}
library(tidyverse)
library(janitor)
library(forcats)
```


## Import the data

This was filtered previously.

```{r}
fa_travis <- read_rds("data-processed/fa_travis.rds")
```

## Build LILA and LA sets

- We work with Low Income, Low Access first, and include all other relevant columns. We do this by dropping the LA columns.
- We build a Low Access fram with just the LA columns and the Census Tract to join with later.

```{r}
# Creates LILA df by dropping the LA "Low Access" columns
fa_travis_lila <- fa_travis %>% 
  select(-starts_with("LA"))

# Create a LA set for later
fa_travis_la <- fa_travis %>% 
  select(CensusTract, Urban, LATracts_half, LATracts1, LATracts10, LATracts20)
```

## Check outer grocery ranges

We can simplify the data if there are no tracts with food deserts further than 10-miles range.

### LA check on grocery range

This shows there are no tracts with grocery store further away than 10 miles. This means we only really need LATracts_half, LATracts1 since no tracts have anything further.

```{r}
fa_travis_la %>%
  filter(LATracts10 == 1)
```

### Do we have rural coutines in LILA?

Since there is a more complicated rural component to the LILA variables, we will check if any of our rural counties have food deserts outside the 10-mile range.

Are there non-Urban counties?

```{r}
fa_travis_lila %>% 
  tabyl(Urban)
```

Yes ... 22 of them.
 
### Are any rural counties LILA food deserts?

```{r}
# filter for rural counties
fa_travis_lila_rural <- fa_travis_lila %>% 
  filter(Urban == 0)

fa_travis_lila_rural %>% 
  tabyl(Urban, LILATracts_1And10)

fa_travis_lila_rural %>% 
  tabyl(Urban, LILATracts_1And20)
```

There are 22 rural tracts, but none are food deserts at 10+ or 20+ miles, so we can exclude them.

## Build Food Access columns

### Reorder LILA columns for clarity

In help to understand these food desert columns, I'm putting them in order in the dataframe:

- LILATracts_halfAnd10: Means food access is between 1/2 mile and 1 mile away in urban areas.
- LILATracts_1And10: Means food access is further than 1 mile away in urban area.
- LILATracts_1And20: Means food access is further than 1 mile away in urban area, and futher than 20 miles in rural areas. We've found above we don't have any in Travis county, so we are dropping it.
- LILATracts_Vehicle: Is a confusing metric outside our story. We will drop it.

```{r}
# reorders columns with LILA at end in order
# this seems overly complicated, but it worked
fa_travis_lila <- fa_travis_lila %>%
  select(-LILATracts_halfAnd10, everything()) %>%
  select(-LILATracts_1And10, everything()) %>%
  select(-LILATracts_1And20, everything()) %>%
  select(-LILATracts_Vehicle, everything())

# dropping 20-mile rural and Vehicle field
fa_travis_lila <- fa_travis_lila %>% 
  select(-LILATracts_1And20, -LILATracts_Vehicle)

```

### Create column to show LILA food access for tracts

How to rate the tracts. Remember we already checked the 10+, 20+ for rural tracts.

- 00 -- Not a food desert. THIS SEEMS OFF. ESPECIALLY WITH LILA.
- 10 -- Supermarket more than 1/2 mile away (but within a mile?)
- 11 -- Supermarket more than 1 mile away.

Low-income census tracts where a significant number or share of residents is more than 1/2 mile (urban) or 10 miles (rural) from the nearest supermarket.

Low-income census tracts where a significant number or share of residents is more than 1 mile (urban) or 10 miles (rural) from the nearest supermarket.


```{r}
fa_travis_lila <- fa_travis_lila %>% 
  mutate(Limited_Income_Limited_Access = 
            case_when(
              (LILATracts_halfAnd10 == 0 & LILATracts_1And10 == 0) ~ "Good access or rural",
              (LILATracts_halfAnd10 == 1 & LILATracts_1And10 == 0) ~ "Between 1/2 and 1 mile",
              (LILATracts_halfAnd10 == 1 & LILATracts_1And10 == 1) ~ "More than 1 mile"
            )
        )

# peeks at the results for veracity
fa_travis_lila %>%
  select(CensusTract,
         LILATracts_halfAnd10,
         LILATracts_1And10,
         Limited_Income_Limited_Access
  ) %>% head(10)

```


### Create column for LA food access for tracts

We are picking up the fa_travis_la dataframe from above and creating a similar Food Access column

```{r}
fa_travis_la <- fa_travis_la %>% 
  mutate(Limited_Access = 
            case_when(
              (LATracts_half == 0 & LATracts1 == 0) ~ "Good access or rural",
              (LATracts_half == 1 & LATracts1 == 0) ~ "Between 1/2 and 1 mile",
              (LATracts_half == 1 & LATracts1 == 1) ~ "More than 1 mile"
            )
        )

# peeking to check
# we need 20 to check all three versions
fa_travis_la %>% 
  select(CensusTract, LATracts_half, LATracts1, Limited_Access) %>% 
  head(20)

```








## Join LA to LILA

We'll create a new dataframe that starts with LILA and adds the `Limited Access` column from the LA dataframe.

```{r}
# Drop unneeded columns from la
fa_travis_la <- fa_travis_la %>% 
  select(CensusTract, LATracts_half, LATracts1, Limited_Access)

# join the dataframes
tfa <- left_join(fa_travis_lila, fa_travis_la, by="CensusTract")

# peek at names to check join
tfa %>% names()

```


### Set the food access columns as factors

This sets these columns as factors and sets the level from best access to worst. This will be useful later with mapping in R.

```{r}
tfa <- tfa %>% 
  mutate(
    Limited_Income_Limited_Access = Limited_Income_Limited_Access %>%
      as_factor() %>% 
      fct_relevel(
        "Good access or rural",
        "Between 1/2 and 1 mile",
        "More than 1 mile"
        ),
    Limited_Access = Limited_Access %>%
      as_factor() %>% 
      fct_relevel(
        "Good access or rural",
        "Between 1/2 and 1 mile",
        "More than 1 mile"
        )
    )

## Peek at the levels
tfa$Limited_Access %>% levels()
```

## Export for QGIS

### Remove columns

We'll select only the columns we need in the final dataset.

```{r}
tfa <- tfa %>% 
  select(
    CensusTract,
    Urban,
    LowIncomeTracts,
    POP2010,
    PovertyRate,
    MedianFamilyIncome,
    Limited_Access,
    Limited_Income_Limited_Access
    )

# peek at names
tfa %>% names()

```

### Fix boolean values

There are some true/false vales represented as 1 and 0. Let's change them.

```{r}
# run thss only once
tfa <- tfa %>% 
  mutate(
    Urban = 
      case_when(
        Urban == 0 ~ "No",
        Urban == 1 ~ "Yes"
      ),
    LowIncomeTracts =
      case_when(
        LowIncomeTracts == 0 ~ "No",
        LowIncomeTracts == 1 ~ "Yes"
      )
  )

```


### Do the exports

Exports to csv to then join in QGIS.

Also exports as .rds to build maps in R.

```{r}
tfa %>% write_rds("data-processed/tfa.rds")
tfa %>% write_csv("data-processed/tfa.csv")
```
